Time: 23 ms (74.77%), Space: 13.1 MB (74.99%) - LeetHub

+++ 0115-distinct-subsequences/0115-distinct-subsequences.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int memo[1005][1005];
-    long long solve(string& s, string& t, int i, int j) {
+    int solve(string& s, string& t, int i, int j) {
         if (j == 0) return 1;
         if (i == 0) return 0;
         if (memo[i][j] != -1) return memo[i][j];
@@ -12,7 +12,7 @@ public:
             return memo[i][j] = solve(s, t, i - 1, j);
         }
     }
-    
+
     int numDistinct(string s, string t) {
         memset(memo, -1, sizeof(memo));
         return solve(s, t, s.length(), t.length());
